Add some comments. Removed 64-bit warning.
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@13257 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/ri-util/size.c
@@ -229,7 +229,7 @@ int type_memory_size(type t)
 represented by a basic, except for a varying size string (formal
 parameter). */
 
-int SizeOfElements(basic b)
+intptr_t SizeOfElements(basic b)
 {
   int e = -1;
